Support selenium grid for discord bot #1559
