+++ lncrawl/core/exeptions.py
@@ -2,6 +2,7 @@ from urllib.error import URLError
 
 from cloudscraper.exceptions import CloudflareException
 from requests.exceptions import RequestException
+from urllib3.exceptions import HTTPError
 
 
 class LNException(Exception):
@@ -14,6 +15,7 @@ class FallbackToBrowser(Exception):
 
 ScraperErrorGroup = (
     URLError,
+    HTTPError,
     CloudflareException,
     RequestException,
     FallbackToBrowser,
